Fixed navigation issue and merged in view to correct location
