+++ src/screens/renterFlatFindScreens/SubScreens/FlatListSubScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ScrollView, StyleSheet, SafeAreaView} from 'react-native';
+import {ScrollView, StyleSheet, SafeAreaView, Pressable} from 'react-native';
 
 // Components 🪢
 import ListViewFlatCard from '@Components/cards/ListViewFlatCard';
@@ -10,16 +10,19 @@ const FlatListSubScreen = ({flats, navigation}: any) => {
       <SafeAreaView>
         {flats.map((el: any, index: number) => {
           return (
-            <ListViewFlatCard
-              navigation={navigation}
-              key={index + 1}
-              match={el?.matchP}
-              flatId={el.flatId}
-              district={el.district}
-              price={el.price}
-              images={el.images}
-              likedUsers={el.likedUsers}
-            />
+            <Pressable
+              onPress={() => navigation.navigate('flatProfile')}
+              key={index}>
+              <ListViewFlatCard
+                navigation={navigation}
+                match={el?.matchP}
+                flatId={el.flatId}
+                district={el.district}
+                price={el.price}
+                images={el.images}
+                likedUsers={el.likedUsers}
+              />
+            </Pressable>
           );
         })}
       </SafeAreaView>
